requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.0.1=pypi_0
aiofiles=23.2.1=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=24.3.0=pypi_0
av=14.0.1=pypi_0
blas=1.0=mkl
blobfile=3.0.0=pypi_0
brotli-python=1.0.9=py310h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.11.26=h06a4308_0
certifi=2024.12.14=py310h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.8=pypi_0
contourpy=1.3.1=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.6.77=0
cuda-runtime=12.1.0=0
cuda-version=12.6=3
cycler=0.12.1=pypi_0
datasets=3.1.0=pypi_0
deepspeed=0.15.4=pypi_0
dill=0.3.8=pypi_0
docstring-parser=0.16=pypi_0
einops=0.8.0=pypi_0
exceptiongroup=1.2.2=pypi_0
fairscale=0.4.13=pypi_0
fastapi=0.115.6=pypi_0
ffmpeg=4.3=hf484d3e_0
ffmpy=0.5.0=pypi_0
filelock=3.13.1=py310h06a4308_0
fire=0.7.0=pypi_0
fonttools=4.55.3=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.5.0=pypi_0
fsspec=2024.9.0=pypi_0
giflib=5.2.2=h5eee18b_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
gradio=4.44.1=pypi_0
gradio-client=1.3.0=pypi_0
h11=0.14.0=pypi_0
hjson=3.1.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.27.1=pypi_0
idna=3.7=py310h06a4308_0
importlib-resources=6.5.2=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jieba=0.42.1=pypi_0
jinja2=3.1.4=py310h06a4308_1
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
kiwisolver=1.4.8=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.11.1.6=0
libcurand=10.3.7.77=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
llama3=0.0.1=dev_0
llamafactory=0.9.2.dev0=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lxml=5.3.0=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py310h5eee18b_0
matplotlib=3.10.0=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.11=py310h5eee18b_0
mkl_random=1.2.8=py310h1128e8f_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py310h06a4308_0
ninja=1.11.1.3=pypi_0
nltk=3.9.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-ml-py=12.560.30=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.15=h5eee18b_0
orjson=3.10.13=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
peft=0.12.0=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py310h06a4308_0
propcache=0.2.1=pypi_0
protobuf=5.29.2=pypi_0
psutil=6.1.1=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=18.1.0=pypi_0
pycryptodomex=3.21.0=pypi_0
pydantic=2.10.4=pypi_0
pydantic-core=2.27.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.20=pypi_0
pytorch=2.5.1=py3.10_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_6
pytorch-mutex=1.0=cuda
pytz=2024.2=pypi_0
pyyaml=6.0.2=py310h5eee18b_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.3=py310h06a4308_1
rich=13.9.4=pypi_0
rouge-chinese=1.0.3=pypi_0
ruff=0.8.6=pypi_0
safetensors=0.5.1=pypi_0
scipy=1.15.0=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=75.1.0=py310h06a4308_0
shellingham=1.5.4=pypi_0
shtab=1.7.1=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
sse-starlette=2.2.1=pypi_0
starlette=0.41.3=pypi_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
termcolor=2.5.0=pypi_0
tiktoken=0.4.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.20.3=pypi_0
tomlkit=0.12.0=pypi_0
torchaudio=2.5.1=py310_cu121
torchtriton=3.1.0=py310
torchvision=0.20.1=py310_cu121
tqdm=4.67.1=pypi_0
transformers=4.46.1=pypi_0
trl=0.9.6=pypi_0
typer=0.15.1=pypi_0
typing_extensions=4.12.2=py310h06a4308_0
tyro=0.8.14=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=py310h06a4308_0
uvicorn=0.34.0=pypi_0
websockets=12.0=pypi_0
wheel=0.44.0=py310h06a4308_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.18.3=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
